requirements.txt
absl-py==1.0.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
boto3==1.23.3
botocore==1.26.3
branca==0.5.0
CacheControl==0.12.11
cachetools==5.1.0
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
configparser==5.2.0
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
folium==0.12.1.post1
functions==0.7.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
idna==3.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso3166==2.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
mapbox==0.18.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
openpyxl==3.0.9
ortools==9.3.10497
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
polyline==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil
pytz
pytz-deprecation-shim
pyzmq
requests
s3transfer
Send2Trash
six
smmap
soupsieve
SQLAlchemy
stack-data
streamlit==1.3.1
streamlit-folium
terminado
tinycss2
toml
toolz
tornado
traitlets
tzdata
tzlocal
uritemplate
urllib3
validators
watchdog
wcwidth
webencodings
widgetsnbextension
zipp
